Fix another double free
If the editor path given is absolute, the same memory address will be used
within the `editor` item from JSON object `obj` and for `editor_args[0]`.

The address will first be freed when freeing editor_args and then a
second time when deleting `obj` by calling `cJSON_Delete()`.

Fix this by returning a copy of `executable` inside `which` when it's an
absolute path.